Added features
Clicking on a game will now show all the information and screenshots. Clicking on a genre redirects to a certain page, but loading games from this genre does not work yet. Installed packages to make the screenshots appearance better if time permits it. Also changed the appearance of the games on the homepage and the genre page